Move controllers to play.controllers
