--- documentation/manual/working/commonGuide/assets/code/ConfiguredAssets.scala
@@ -1,8 +1,0 @@
-/*
- * Copyright (C) 2009-2017 Lightbend Inc. <https://www.lightbend.com>
- */
-package configured.assets
-
-package object controllers {
-  type Assets = _root_.controllers.Assets
-}
