+++ documentation/manual/working/commonGuide/assets/code/assets/CommonAssets.scala
@@ -1,0 +1,14 @@
+
+package common.assets {
+  package object assets {
+    type Assets = play.controllers.assets.Assets
+    val Assets = play.controllers.assets.Assets
+  }
+}
+
+package configured.assets {
+  package object assets {
+    type Assets = play.controllers.assets.Assets
+    val Assets = play.controllers.assets.Assets
+  }
+}
