+++ documentation/manual/working/javaGuide/main/http/code/javaguide/Default.scala
@@ -1,0 +1,6 @@
+package javaguide.http.routing.defaultcontroller
+
+package object util {
+  type Default = play.controllers.util.Default
+  val Default = play.controllers.util.Default
+}
