+++ documentation/manual/working/scalaGuide/main/http/code/scalaguide/Default.scala
@@ -1,0 +1,6 @@
+package scalaguide.http.routing.defaultcontroller
+
+package object util {
+  type Default = play.controllers.util.Default
+  val Default = play.controllers.util.Default
+}
