+++ framework/src/play/src/main/resources/reference.conf
@@ -737,7 +737,7 @@ play {
   filters {
     # List of enabled filters as fully qualified class names
     # enabled = []
-    
+
     # List of disabled filters as fully qualified class names
     disabled = []
   }
@@ -763,7 +763,7 @@ play {
     # The enabled modules that should be automatically loaded.
     enabled += "play.api.inject.BuiltinModule"
     enabled += "play.api.i18n.I18nModule"
-    enabled += "controllers.AssetsModule"
+    enabled += "play.controllers.assets.AssetsModule"
 
     # A way to disable modules that are automatically enabled
     disabled = []
