+++ framework/src/sbt-plugin/src/sbt-test/play-sbt-plugin/dev-mode/conf/routes
@@ -1,2 +1,2 @@
 GET        /                    controllers.Application.index
-GET        /assets/*file        controllers.Assets.versioned(path = "/public", file: Asset)
+GET        /assets/*file        play.controllers.assets.Assets.versioned(path = "/public", file: Asset)
