+++ framework/src/sbt-plugin/src/sbt-test/play-sbt-plugin/distribution/app/views/main.scala.html
@@ -5,9 +5,9 @@
 <html>
     <head>
         <title>@title</title>
-        <link rel="stylesheet" media="screen" href="@routes.Assets.at("stylesheets/main.css")">
-        <link rel="shortcut icon" type="image/png" href="@routes.Assets.at("images/favicon.png")">
-        <script src="@routes.Assets.at("javascripts/jquery-2.2.0.min.js")" type="text/javascript"></script>
+        <link rel="stylesheet" media="screen" href="@play.controllers.assets.routes.Assets.at("stylesheets/main.css")">
+        <link rel="shortcut icon" type="image/png" href="@play.controllers.assets.routes.Assets.at("images/favicon.png")">
+        <script src="@play.controllers.assets.routes.Assets.at("javascripts/jquery-2.2.0.min.js")" type="text/javascript"></script>
     </head>
     <body>
         @content
