+++ framework/src/sbt-plugin/src/sbt-test/routes-compiler-plugin/routes-compilation/conf/routes
@@ -67,7 +67,7 @@ GET         /routestest             controllers.Application.routetest(yield)
 # Test for default values for scala keywords
 GET         /routesdefault          controllers.Application.routedefault(type ?= "x")
 
-GET         /public/*file           controllers.Assets.versioned(path="/public", file: Asset)
+GET         /public/*file           play.controllers.assets.Assets.versioned(path="/public", file: Asset)
 
 # This triggers a string interpolation warning, since there is an identifier called "routes" in scope, and it
 # generates a non interpolated string containing $routes. As does this comment.
